+++ sportspress-player-registration/uninstall.php
@@ -12,10 +12,15 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
-// Remove plugin options
-delete_option( 'spr_auto_create' );
-delete_option( 'spr_auto_update' );
-delete_option( 'spr_auto_role' );
-delete_option( 'spr_player_role' );
-delete_option( 'spr_auto_season' );
-delete_option( 'spr_db_version' );
+// Check if parent plugin wants data removed
+if ( get_option( 'spat_remove_data_on_uninstall', '0' ) === '1' ) {
+
+	// Remove plugin options
+	delete_option( 'spr_auto_create' );
+	delete_option( 'spr_auto_update' );
+	delete_option( 'spr_auto_role' );
+	delete_option( 'spr_player_role' );
+	delete_option( 'spr_auto_season' );
+	delete_option( 'spr_db_version' );
+	delete_option( 'spr_registration_keyword' );
+}
